requirements.txt
django-constance[database]
django-cors-headers
django-environ
django-extensions
django-probes
django-debug-toolbar
djangorestframework
psycopg2
python-dateutil
pool-data-fetcher